+++ tests/suites/kafka_sanity/resources/alpine.yaml
